[Refactoring] Default port manager functions now use PortManager__Generic and LocalOperations
This patch deletes a duplication of port manager code.

Now utils.reserve_port and utils.release_port works through _old_port_manager - it is a global instance of PortManager__Generic that uses a global instance of LocalOperations.

This commit is a part of work for #247.
